+++ psycop/common/model_training_v2/hyperparameter_suggester/suggesters/logistic_regression_suggester.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Any
+from typing import Any, Mapping, Sequence
 
 import optuna
 
@@ -8,7 +8,7 @@ from psycop.common.model_training_v2.hyperparameter_suggester.suggesters.base_su
     Suggester,
 )
 
-FloatSpaceT = list[float | bool]
+FloatSpaceT = Mapping[str, float | bool]
 
 
 @dataclass(frozen=True)
@@ -25,6 +25,13 @@ class FloatSpace:
             log=self.logarithmic,
         )
 
+@dataclass(frozen=True)
+class CategoricalSpace:
+    choices: Sequence[str]
+
+    def suggest(self, trial: optuna.Trial, name: str) -> Any:
+        return trial.suggest_categorical(name=name, choices=self.choices)
+
 
 @BaselineRegistry.estimator_steps.register("mock_suggester")
 class MockSuggester(Suggester):
@@ -37,35 +44,3 @@ class MockSuggester(Suggester):
         return {"mock_value": self.value.suggest(trial, "mock_suggester")}
 
 
-@BaselineRegistry.estimator_steps.register("logistic_regression_suggester")
-class LogisticRegressionSuggester(Suggester):
-    # TODO: #424 Refactor so suggesters are co-located with their corresponding model steps.
-    # Perhaps that makes it hard to decorate, though?
-
-    # TODO: Can suggesters take a mapping for each argument?
-    # E.g. C can be a mapping which must contain low, high and log? How do we type-hint that? A nameddict?
-
-    def __init__(
-        self,
-        C_low: float,
-        C_high: float,
-        C_log: bool,
-        l1_ratio_low: float,
-        l1_ratio_high: float,
-        l1_ratio_log: bool,
-    ):
-        self.C = FloatSpace(low=C_low, high=C_high, logarithmic=C_log)
-        self.l1_ratio = FloatSpace(
-            low=l1_ratio_low,
-            high=l1_ratio_high,
-            logarithmic=l1_ratio_log,
-        )
-
-    def suggest_hyperparameters(self, trial: optuna.Trial) -> dict[str, Any]:
-        return {
-            "logistic_regression": {
-                "@estimator_steps": "logistic_regression",
-                "C": self.C.suggest(trial, "C"),
-                "l1_ratio": self.l1_ratio.suggest(trial, "l1_ratio"),
-            },
-        }
